Add a standalone autodiscover command & refactor some code
Initially, I just wanted to add a stand alone "autodiscover" command. The intention is that you can quickly throw a domain at it, and it'll tell you what the autodiscover URL is. Ruler does some smarter lookups than just looking up the "autodiscover.<domain>" record and I wanted access to that.

I then noticed, that for domains that aren't obviously Office 365 hosted, ruler would tell me through the HTTP 4xx status whether they were. So I added a kludge of a check for that.

In doing so, I had to trace a lot of code, and ended up starting a bit of a refactor of the connect() method in ruler.go. It can be taken further, but I had limited time.

The next step should be to remove some of the now redundant code in mapi.Init() since some of the parameters passed are now in config. Also to remove the "Config" utils.datatype as it doesn't look used. But I don't want to break anything yet.