Updates the search to use the `PR_SEARCH_ALL_INDEXED_PROPS_W` property.
All indexed fields are searched by default, instead of the original
default of searching the body. This means there is now a new flag
`--body` which allows for specifying that the search should happen in
the body. The existing `--subject` functions as before.

_NB_ the `PR_SEARCH_ALL_INDEXED_PROPS_W` property is only compatible
with Exchange 2010 and up. This shouldn't be a deal breaker in most
cases. If errors occur, simply switch to `--body` or `--subject`

Ref:
[https://blogs.msdn.microsoft.com/stephen_griffin/2013/11/08/indexed-search-with-mapi-and-exchange/](https://blogs.msdn.microsoft.com/stephen_griffin/2013/11/08/indexed-search-with-mapi-and-exchange/)